add lua-resty-socket
Hi there,

I'm not sure if this module belongs to utilities or networking, even less if it belongs to this list at all, actually. I have seen several libraries trying to compensate for the lack of support of cosockets in OpenResty's `init` context, and I myself need it for lua-cassandra. pgmoon provided a LuaSocket compatibility module for plain Lua (although not for the `init` context and somewhat incomplete). Hence, I wrote this lua-resty-socket a few months ago as an iteration over it. It comes very handy when one wants to write his or her lua-resty module as if it were always being used in OpenResty, but with automatic support for the `init` context **and** plain Lua (using LuaSocket/LuaSec).

Let me know how you feel about adding it to this list or not.
Cheers